Move PortConverter to HW
