+++ lib/Support/CMakeLists.txt
@@ -19,13 +19,9 @@ add_circt_library(CIRCTSupport
   PrettyPrinterHelpers.cpp
   SymCache.cpp
   ValueMapper.cpp
-  PortConverter.cpp
   "${VERSION_CPP}"
 
   ADDITIONAL_HEADER_DIRS
-
-  LINK_LIBS PUBLIC
-  CIRCTHW
 
   MLIRIR
   MLIRTransforms
